started tile stuff and moved Cell out of Gameboard.ts
